Make changelog truncation configurable (#133)
This PR makes changelog truncation a store-level config, rather than something we implicitly derive/guess at the user intention based on the topic configs. This solves a few things that have been on my mind for a while

1. Don't accidentally wipe out users don't when they weren't expecting it 💀
2. Make it a top-level config so it's easier to notice for people who do want it
3. Since we currently aren't compatible with the source-topic optimization or disabling the changelog, we can point users who are concerned about changelog topic space/growth to use changelog truncation instead
4. Best of all, no need to make so many admin calls during state store initialization (with this and Cache partition counts for sub-partitioner and set definitions/naming conventions #130, we now have no superflous admin calls)

+++ kafka-client/src/main/java/dev/responsive/kafka/store/CommitBuffer.java
@@ -27,7 +27,6 @@ import dev.responsive.kafka.clients.SharedClients;
 import dev.responsive.kafka.config.ResponsiveConfig;
 import dev.responsive.model.Result;
 import dev.responsive.utils.Iterators;
-import dev.responsive.utils.StoreUtil;
 import dev.responsive.utils.TableName;
 import java.time.Duration;
 import java.time.Instant;
@@ -45,6 +44,7 @@ import org.apache.kafka.clients.admin.Admin;
 import org.apache.kafka.clients.admin.RecordsToDelete;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.common.TopicPartition;
+import org.apache.kafka.common.errors.PolicyViolationException;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.errors.ProcessorStateException;
@@ -75,19 +75,21 @@ class CommitBuffer<K, S extends RemoteSchema<K>> implements RecordBatchingStateR
   private Instant lastFlush;
   private WriterFactory<K> writerFactory;
 
+  // flag to skip further truncation attempts when the changelog is set to 'compact' only
+  private boolean isNotDeleteEnabled = false;
+
   static <K, S extends RemoteSchema<K>> CommitBuffer<K, S> from(
       final SharedClients clients,
       final TableName tableName,
       final TopicPartition changelog,
       final S schema,
       final KeySpec<K> keySpec,
+      final boolean truncateChangelog,
       final SubPartitioner partitioner,
       final ResponsiveConfig config
   ) {
     final var admin = clients.admin;
     final var cassandraClient = clients.cassandraClient;
-    final boolean truncate = StoreUtil.shouldTruncateChangelog(
-        changelog.topic(), admin, config.originals());
 
     return new CommitBuffer<>(
         cassandraClient,
@@ -96,7 +98,7 @@ class CommitBuffer<K, S extends RemoteSchema<K>> implements RecordBatchingStateR
         admin,
         schema,
         keySpec,
-        truncate,
+        truncateChangelog,
         FlushTriggers.fromConfig(config),
         partitioner
     );
@@ -402,13 +404,19 @@ class CommitBuffer<K, S extends RemoteSchema<K>> implements RecordBatchingStateR
   }
 
   private void maybeTruncateChangelog(final long offset) {
-    if (truncateChangelog) {
+    if (truncateChangelog && !isNotDeleteEnabled) {
       try {
         admin.deleteRecords(Map.of(changelog, RecordsToDelete.beforeOffset(offset))).all().get();
         log.info("Truncated changelog topic {} before offset {}", changelog, offset);
       } catch (final ExecutionException e) {
-        log.warn("Could not truncate changelog topic-partition {}.", changelog, e);
+        log.warn("Could not truncate changelog topic-partition " + changelog, e);
+        if (e.getCause() instanceof PolicyViolationException) {
+          log.info("Disabling further changelog truncation attempts due to topic configuration "
+                       + "being incompatible with deleteRecord requests", e);
+          isNotDeleteEnabled = true;
+        }
       } catch (final InterruptedException e) {
+        log.error("Interrupted while truncating the changelog " + changelog, e);
         throw new ProcessorStateException("Interrupted while truncating " + changelog, e);
       }
     }
